Sécurité modif url

+++ src/Controller/AnnoncesController.php
@@ -22,7 +22,7 @@ class AnnoncesController extends AbstractController
     public function index(AnnonceRepository $repo): Response
     {
         //Récupére toutes les annonces
-        $annonces = $repo->findby([],['date' => 'desc']);
+        $annonces = $repo->findby([], ['date' => 'desc']);
 
         return $this->render('annonces/index.html.twig', [
             'annonces' => $annonces,
@@ -111,14 +111,19 @@ class AnnoncesController extends AbstractController
     {
         //Récupére l'annonce grace à l'ID (GET)
         $annonce = $repo->find($id);
-
         //Récupére les images correspondantes à l'ID de l'annonce
         $images = $repo2->findBy(['annonce' => $id]);
 
-        return $this->render('annonces/show.html.twig', [
-            'annonce' => $annonce,
-            'images' => $images,
-        ]);
+        if(!$annonce) {
+            $this->addFlash('danger', 'Cet annonce n\'existe pas !');
+            return $this->redirectToRoute('app_annonces');
+        }else
+        {
+            return $this->render('annonces/show.html.twig', [
+                'annonce' => $annonce,
+                'images' => $images,
+            ]);
+        }
     }
 
 
@@ -127,21 +132,46 @@ class AnnoncesController extends AbstractController
     {
         $user = $this->getUser();
 
-        //Récupére les annonces correspondant à l'utilisateur
-        $annoncesByUsers = $repo->findBy(['user' => $user], ['id' => 'desc']);
+        if (!$user) {
+            $this->addFlash('danger', 'Vous devez être connecté !');
+            return $this->redirectToRoute('app_annonces');
+        }
+        else 
+        {
+            //Récupére les annonces correspondant à l'utilisateur
+            $annoncesByUsers = $repo->findBy(['user' => $user], ['id' => 'desc']);
 
-        return $this->render('user/annonces.html.twig', [
-            'annoncesByUsers' => $annoncesByUsers,
-        ]);
+            return $this->render('user/annonces.html.twig', [
+                'annoncesByUsers' => $annoncesByUsers,
+            ]);
+        }
     }
 
 
     #[Route('/annonces/{id}/edit', name: 'edit_annonce')]
-    public function editAnnonce(AnnonceRepository $repo, ImageRepository $repo2, Request $request,$id,EntityManagerInterface $entityManager)
+    public function editAnnonce(AnnonceRepository $repo, ImageRepository $repo2, Request $request, $id, EntityManagerInterface $entityManager)
     {
         //Récupére l'annonce concerné
         $annonce = $repo->find($id);
-        
+        //récupére l'email du  propriètaire de l'annonce
+        $user = $annonce->getUser();
+        //récupére l'email de l'utilisateur
+        $userCo = $this->getUser();
+
+        if(!$userCo)
+        {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit de modifier cet annonce !');
+            return $this->redirectToRoute('app_annonces');
+        }
+
+        if($user != $userCo)
+        {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit de modifier cet annonce !');
+            return $this->redirectToRoute('app_annonces');
+        }
+        else
+        {
+
         //Récupére les images correspondantes 
         $imagesByAnnonces = $repo2->findBy(['annonce' => $id]);
 
@@ -151,45 +181,41 @@ class AnnoncesController extends AbstractController
         //Analyse de la requete
         $formEdit->handleRequest($request);
 
-        if ($formEdit->isSubmitted() && $formEdit->isValid())
-        {
+        if ($formEdit->isSubmitted() && $formEdit->isValid()) {
             //Récupere l'image cover transmise
             $annonceFile = $formEdit->get('imageCover')->getData();
 
-            if($annonceFile) 
-            {
+            if ($annonceFile) {
                 //Récupére le chemin
                 $cheminImageCoverSupp = $this->getParameter('annonce_directory') . '/' . $annonce->getImageCover();
 
                 //Si il existe, on supprime du folder
-                if(file_exists($cheminImageCoverSupp))
-                {
-                   unlink($cheminImageCoverSupp);
+                if (file_exists($cheminImageCoverSupp)) {
+                    unlink($cheminImageCoverSupp);
                 }
 
                 //Génére un nouveau nom de fichier pour l'image de couverture
                 $fichierImageCover = md5(uniqid()) . '.' . $annonceFile->guessExtension();
- 
+
                 //Envoie du fichier dans le folder
                 $annonceFile->move(
                     $this->getParameter('annonce_directory'),
                     $fichierImageCover
                 );
- 
-                $annonce->setImageCover($fichierImageCover);
 
-            }else{}
- 
- 
+                $annonce->setImageCover($fichierImageCover);
+            } else {
+            }
+
+
             //Récupére le(s) image(s) transmise(s)
             $images = $formEdit->get('images')->getData();
- 
+
             //Boucle sur les images
-            foreach ($images as $image) 
-            {
+            foreach ($images as $image) {
                 //Génére un nouveau nom de fichier pour les images
                 $fichierImages = md5(uniqid()) . '.' . $image->guessExtension();
- 
+
                 //Envoie des images dans le folder
                 $image->move(
                     $this->getParameter('annonce_directory'),
@@ -200,9 +226,8 @@ class AnnoncesController extends AbstractController
                 $img = new Image();
 
                 $img->setImage($fichierImages)
-                    ->setAnnonce($annonce)
-                ;
-                
+                    ->setAnnonce($annonce);
+
                 //Enregistrement en  BDD
                 $entityManager->persist($img);
             }
@@ -211,23 +236,23 @@ class AnnoncesController extends AbstractController
             $user = $this->getUser();
 
             $annonce->setDate(new \DateTime())
-                    ->setUser($user)
-            ;
-                    
+                ->setUser($user);
+
             $entityManager->persist($annonce);
             $entityManager->flush();
 
             $this->addFlash('success', 'Votre annonce a bien été modifié !');
             return $this->redirectToRoute('app_annonces');
+        } else {
+        }
 
-        }else{}
+        }
 
-        return $this->render('annonces/edit.html.twig',[
+        return $this->render('annonces/edit.html.twig', [
             'annonce' => $annonce,
             'imagesByAnnonces' => $imagesByAnnonces,
             'formEdit' => $formEdit->createView()
         ]);
-
     }
 
 
@@ -237,13 +262,24 @@ class AnnoncesController extends AbstractController
     {
         //Récupére l'image concerné
         $images = $RepoImage->find($id);
+        //récupére le propiétaire de limage
+        $userImage = $images->getAnnonce()->getUser();
+        //récupére l'utilisateur
+        $user = $this->getUser();
+
+        if($userImage != $user)
+        {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit de supprimer cet image !');
+            return $this->redirectToRoute('app_annonces');
+        }
+        else
+        {
 
         //Récupére le chemin
         $cheminImage = $this->getParameter('annonce_directory') . '/' . $images->getImage();
 
         //Si il existe, on supprime du folder
-        if(file_exists($cheminImage))
-        {
+        if (file_exists($cheminImage)) {
             unlink($cheminImage);
         }
 
@@ -255,18 +291,32 @@ class AnnoncesController extends AbstractController
         return $this->redirectToRoute('edit_annonce', [
             'id' => $images->getAnnonce()->getId()
         ]);
-        
+
+    }
     }
 
 
 
     #[Route('/annonces/{id}/supp', name: 'supp_annonce')]
-    public function supprimerAnnonce(AnnonceRepository $repoAnnonce, ImageRepository $repoImage ,$id, EntityManagerInterface $entityManager)
+    public function supprimerAnnonce(AnnonceRepository $repoAnnonce, ImageRepository $repoImage, $id, EntityManagerInterface $entityManager)
     {
         //Récupére l'annonce concerné
         $annonce = $repoAnnonce->find($id);
         //Récupére les images lié a cet annonce
         $images = $repoImage->findBy(['annonce' => $id]);
+        //Récupére le propriétaire de l'annonce
+        $userAnnonce = $annonce->getUser();
+        //Récupére l'utilisateur
+        $user = $this->getUser();
+
+
+        if($userAnnonce != $user)
+        {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit de supprimer cet image !');
+            return $this->redirectToRoute('app_annonces');
+        }
+        else
+        {
 
         //Récupére l'image de couverture de l'annonce
         $imageCover = $annonce->getImageCover();
@@ -274,23 +324,20 @@ class AnnoncesController extends AbstractController
         $cheminCoverImage = $this->getParameter('annonce_directory') . '/' . $imageCover;
 
         //Si il existe, on supprime du folder
-        if(file_exists($cheminCoverImage))
-        {
+        if (file_exists($cheminCoverImage)) {
             unlink($cheminCoverImage);
         }
 
         //Récupére tous les chemins des image(s) correspondantes
-        foreach($images as $image)
-        {
+        foreach ($images as $image) {
             $cheminImage = $this->getParameter('annonce_directory') . '/' . $image->getImage();
 
             //Si existe, on supprime du folder
-            if(file_exists($cheminImage))
-            {
+            if (file_exists($cheminImage)) {
                 unlink($cheminImage);
             }
         }
-  
+
         //Supprimer l'annonce de la BDD
         $entityManager->remove($annonce);
         $entityManager->flush();
@@ -300,5 +347,7 @@ class AnnoncesController extends AbstractController
         return $this->redirectToRoute('annoncesByUser');
 
     }
+    }
 }
+
